Moving actions to DocumentListActions
